requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
jinja2==3.1.2
openai>=1.6.1,<2.0.0
tiktoken>=0.5.2,<0.6.0
langchain==0.0.339
langchain-openai==0.0.2.post1
chromadb==0.4.18
PyPDF2==3.0.1
python-docx==0.8.11
docx2txt==0.8
unstructured==0.10.30
openpyxl==3.1.2 